MWPW-168739 - Modularize transition screen code
