+++ unitylibs/core/workflow/workflow-acrobat/upload-handler.js
@@ -139,7 +139,9 @@ export default class UploadHandler {
           await this.actionBinder.dispatchErrorToast('verb_upload_error_generic', 500, `Unexpected response from finalize call: ${assetData.id}, ${JSON.stringify(finalizeJson)}`);
           return false;
         }
-        await this.actionBinder.showSplashScreen();
+        const { default: TransitionScreen } = await import(`${getUnityLibs()}/scripts/transition-screen.js`);
+        this.transitionScreen = new TransitionScreen(this.actionBinder.transitionScreen.splashScreenEl, this.actionBinder.initActionListeners, this.actionBinder.LOADER_LIMIT, this.actionBinder.workflowCfg);
+        await this.transitionScreen.showSplashScreen();
         await this.actionBinder.dispatchErrorToast('verb_upload_error_generic', 500, `Unexpected response from finalize call: ${assetData.id}, ${JSON.stringify(finalizeJson)}`);
         this.actionBinder.operations = [];
         return false;
@@ -149,7 +151,9 @@ export default class UploadHandler {
         await this.actionBinder.dispatchErrorToast('verb_upload_error_generic', e.status || 500, `Exception thrown when verifying content: ${e.message}, ${assetData.id}`, false, e.showError);
         return false;
       }
-      await this.actionBinder.showSplashScreen();
+      const { default: TransitionScreen } = await import(`${getUnityLibs()}/scripts/transition-screen.js`);
+      this.transitionScreen = new TransitionScreen(this.actionBinder.transitionScreen.splashScreenEl, this.actionBinder.initActionListeners, this.actionBinder.LOADER_LIMIT, this.actionBinder.workflowCfg);
+      await this.transitionScreen.showSplashScreen();
       await this.actionBinder.dispatchErrorToast('verb_upload_error_generic', e.status || 500, `Exception thrown when verifying content: ${e.message}, ${assetData.id}`, false, e.showError);
       this.actionBinder.operations = [];
       return false;
@@ -170,7 +174,9 @@ export default class UploadHandler {
           if (this.actionBinder?.limits?.pageLimit?.maxNumPages
             && metadata.numPages > this.actionBinder.limits.pageLimit.maxNumPages
           ) {
-            await this.actionBinder.showSplashScreen();
+            const { default: TransitionScreen } = await import(`${getUnityLibs()}/scripts/transition-screen.js`);
+            this.transitionScreen = new TransitionScreen(this.actionBinder.transitionScreen.splashScreenEl, this.actionBinder.initActionListeners, this.actionBinder.LOADER_LIMIT, this.actionBinder.workflowCfg);
+            await this.transitionScreen.showSplashScreen();
             await this.actionBinder.dispatchErrorToast('verb_upload_error_max_page_count');
             resolve(true);
             return;
@@ -178,7 +184,9 @@ export default class UploadHandler {
           if (this.actionBinder?.limits?.pageLimit?.minNumPages
             && metadata.numPages < this.actionBinder.limits.pageLimit.minNumPages
           ) {
-            await this.actionBinder.showSplashScreen();
+            const { default: TransitionScreen } = await import(`${getUnityLibs()}/scripts/transition-screen.js`);
+            this.transitionScreen = new TransitionScreen(this.actionBinder.transitionScreen.splashScreenEl, this.actionBinder.initActionListeners, this.actionBinder.LOADER_LIMIT, this.actionBinder.workflowCfg);
+            await this.transitionScreen.showSplashScreen();
             await this.actionBinder.dispatchErrorToast('verb_upload_error_min_page_count');
             resolve(true);
             return;
@@ -207,7 +215,9 @@ export default class UploadHandler {
         }, totalDuration);
       });
     } catch (e) {
-      await this.actionBinder.showSplashScreen();
+      const { default: TransitionScreen } = await import(`${getUnityLibs()}/scripts/transition-screen.js`);
+      this.transitionScreen = new TransitionScreen(this.actionBinder.transitionScreen.splashScreenEl, this.actionBinder.initActionListeners, this.actionBinder.LOADER_LIMIT, this.actionBinder.workflowCfg);
+      await this.transitionScreen.showSplashScreen();
       await this.actionBinder.dispatchErrorToast('verb_upload_error_generic', e.status || 500, `Exception thrown when verifying PDF page count; ${e.message}`, false, e.showError);
       this.actionBinder.operations = [];
       return false;
@@ -233,7 +243,9 @@ export default class UploadHandler {
 
   async dispatchGenericError(info = null, showError = true) {
     this.actionBinder.operations = [];
-    await this.actionBinder.showSplashScreen();
+    const { default: TransitionScreen } = await import(`${getUnityLibs()}/scripts/transition-screen.js`);
+    this.transitionScreen = new TransitionScreen(this.actionBinder.transitionScreen.splashScreenEl, this.actionBinder.initActionListeners, this.actionBinder.LOADER_LIMIT, this.actionBinder.workflowCfg);
+    await this.transitionScreen.showSplashScreen();
     await this.actionBinder.dispatchErrorToast('verb_upload_error_generic', 500, info, false, showError);
   }
 
@@ -331,8 +343,10 @@ export default class UploadHandler {
   }
 
   async singleFileGuestUpload(file, fileData) {
+    const { default: TransitionScreen } = await import(`${getUnityLibs()}/scripts/transition-screen.js`);
+    this.transitionScreen = new TransitionScreen(this.actionBinder.transitionScreen.splashScreenEl, this.actionBinder.initActionListeners, this.actionBinder.LOADER_LIMIT, this.actionBinder.workflowCfg);
     try {
-      await this.actionBinder.showSplashScreen(true);
+      await this.transitionScreen.showSplashScreen(true);
       if (this.isNonPdf([file])) {
         await this.actionBinder.delay(3000);
         const redirectSuccess = await this.actionBinder.handleRedirect(this.getGuestConnPayload('nonpdf'));
@@ -342,18 +356,20 @@ export default class UploadHandler {
       }
       await this.uploadSingleFile(file, fileData);
     } catch (e) {
-      await this.actionBinder.showSplashScreen();
+      await this.transitionScreen.showSplashScreen();
       this.actionBinder.operations = [];
       await this.handleUploadError(e);
     }
   }
 
   async singleFileUserUpload(file, fileData) {
+    const { default: TransitionScreen } = await import(`${getUnityLibs()}/scripts/transition-screen.js`);
+    this.transitionScreen = new TransitionScreen(this.actionBinder.transitionScreen.splashScreenEl, this.actionBinder.initActionListeners, this.actionBinder.LOADER_LIMIT, this.actionBinder.workflowCfg);
     try {
-      await this.actionBinder.showSplashScreen(true);
+      await this.transitionScreen.showSplashScreen(true);
       await this.uploadSingleFile(file, fileData, this.isNonPdf([file]));
     } catch (e) {
-      await this.actionBinder.showSplashScreen();
+      await this.transitionScreen.showSplashScreen();
       this.actionBinder.operations = [];
       await this.handleUploadError(e);
     }
@@ -384,7 +400,10 @@ export default class UploadHandler {
       return;
     }
     this.actionBinder.LOADER_LIMIT = 75;
-    this.actionBinder.updateProgressBar(this.actionBinder.splashScreenEl, 75);
+
+    const { default: TransitionScreen } = await import(`${getUnityLibs()}/scripts/transition-screen.js`);
+    this.transitionScreen = new TransitionScreen(this.actionBinder.transitionScreen.splashScreenEl, this.actionBinder.initActionListeners, this.actionBinder.LOADER_LIMIT, this.actionBinder.workflowCfg);
+    this.transitionScreen.updateProgressBar(this.actionBinder.transitionScreen.splashScreenEl, 75);
     cOpts = {
       targetProduct: this.actionBinder.workflowCfg.productName,
       assetId: assetDataArray[0].id,
@@ -419,15 +438,17 @@ export default class UploadHandler {
     if (allVerified === 0) return;
     if (files.length !== allVerified) this.actionBinder.multiFileFailure = 'uploaderror';
     this.actionBinder.LOADER_LIMIT = 95;
-    this.actionBinder.updateProgressBar(this.actionBinder.splashScreenEl, 95);
+    this.transitionScreen.updateProgressBar(this.actionBinder.transitionScreen.splashScreenEl, 95);
   }
 
   async multiFileGuestUpload(filesData) {
     try {
-      await this.actionBinder.showSplashScreen(true);
+      const { default: TransitionScreen } = await import(`${getUnityLibs()}/scripts/transition-screen.js`);
+      this.transitionScreen = new TransitionScreen(this.actionBinder.transitionScreen.splashScreenEl, this.actionBinder.initActionListeners, this.actionBinder.LOADER_LIMIT, this.actionBinder.workflowCfg);
+      await this.transitionScreen.showSplashScreen(true);
       await this.actionBinder.delay(3000);
       this.actionBinder.LOADER_LIMIT = 85;
-      this.actionBinder.updateProgressBar(this.actionBinder.splashScreenEl, 85);
+      this.transitionScreen.updateProgressBar(this.actionBinder.transitionScreen.splashScreenEl, 85);
       const redirectSuccess = await this.actionBinder.handleRedirect(this.getGuestConnPayload('multifile'));
       if (!redirectSuccess) return;
       this.actionBinder.redirectWithoutUpload = true;
@@ -439,7 +460,9 @@ export default class UploadHandler {
 
   async multiFileUserUpload(files, filesData) {
     try {
-      await this.actionBinder.showSplashScreen(true);
+      const { default: TransitionScreen } = await import(`${getUnityLibs()}/scripts/transition-screen.js`);
+      this.transitionScreen = new TransitionScreen(this.actionBinder.transitionScreen.splashScreenEl, this.actionBinder.initActionListeners, this.actionBinder.LOADER_LIMIT);
+      await this.transitionScreen.showSplashScreen(true);
       await this.uploadMultiFile(files, filesData);
     } catch (e) {
       await this.dispatchGenericError(`Exception raised when uploading multiple files for a signed-in user; ${e.message}, Files data: ${JSON.stringify(filesData)}`, e.showError);
